fix: cast features to float32 before returning from model_fun
autocast (float16) output was propagating to .numpy() and being saved
to H5, causing dtype=float16 in output features instead of float32.

Add .float() before .cpu() in all model_fun implementations that use
torch.autocast: TorchModel (base), VirchowModel, PhikonModel,
Midnight12kModel, HibouLModel, GigapathModel, GigapathSlideEncoderModel,
and TitanSlideEncoderModel.

+++ mussel/models/base.py
@@ -234,7 +234,7 @@ class TorchModel(Model):
                         x = x.to(memory_format=torch.channels_last)
                     except Exception:
                         pass  # Some tensor shapes may not support channels_last
-                return self.obj(x).cpu()
+                return self.obj(x).float().cpu()
 
         return model_fun
